% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rational.R
\name{rational}
\alias{rational}
\title{Rational Number}
\usage{
rational(num, denom)
}
\arguments{
\item{num}{(integer) numerator}

\item{denom}{(integer) denominator}
}
\description{
Data type and functions to work with rational numbers.
}
\details{
...
}
\examples{
rat <- rational(1:3, 3:5)
rat + 1
}
